+++ src/options/getCommandManagerOptions.ts
@@ -1,10 +1,12 @@
 import path from 'path';
 import type Types from '../../typings';
 
-export default (options: Types.SucroseOptions): Types.CommandManagerOptions => {
+export default (
+  options: Types.SucroseOptions,
+): Types.InteractionCommandManagerOptions => {
   const env = options.env as Types.EnvironmentOptions;
   const logging = options.logging as Types.SucroseLoggerOptions;
-  const directories = options.directories?.commands as Types.CommandDirectories;
+  const directories = options.directories?.interactions.commands as Types.CommandDirectories;
 
   const source = path.join(process.cwd(), env.source);
   const globalCommandsPath = path.join(source, directories.globals);