+++ src/options/getFeaturesOptions.ts
@@ -1,7 +1,7 @@
 import type Types from '../../typings';
 import * as contents from '../contents';
 
-export default (options: Types.SucroseOptions): Types.Features<true> => ({
+export default (options: Types.SucroseOptions): Types.Features => ({
   interactions: {
     autoReply: {
       active: typeof options.features?.interactions?.autoReply?.active === 'boolean' ? options.features.interactions.autoReply.active : true,